agent/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
openai
requests
langgraph
pydantic
sse-starlette
python-dotenv
pytest
